fts: widen FtsReader fields + extract reader/scorers.rs
Phase 2. Split the disjunction scorers (WAND+BMW, MaxScore+BMM,
windowed, exhaustive) with their dispatch and the AND flat-merge
intersection family out of the core impl into reader/scorers.rs as its
own impl FtsReader block. FtsReader's fields and the cross-called
helpers (build_term_cursors, resolve_column_id, the OrAlgo routing
consts/predicates) are widened to pub(super), scoped to reader/;
search_with_algo_for_bench stays pub for the bench harness. Pure move —
FTS oracle/count/phrase + reader-unit + supertable-query suites green.
